requirements.txt
tensorflow==2.8.0
pytesseract
numpy
pillow
opencv-python
streamlit